% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_basis.R
\name{build_basis}
\alias{build_basis}
\title{Setup basis functions.}
\usage{
build_basis(x, disttype = "exp", parameters_values = seq(from = 1, to = 5,
  by = 1), parameters_names = c("rate"), dens = FALSE)
}
\arguments{
\item{x}{numeric vector.}

\item{disttype}{character string; see details.
Default is \code{"exp"} for the exponential distribution.}

\item{parameters_values}{numeric vector or matrix; containing a sequence of
values for the distribution parameters. The no. of columns must be equal to
the no. of parameters. \cr
Default is a sequence from 1 to 5.}

\item{parameters_names}{character vector; containing the names for the
distribution parameters. \cr
Default is \code{"rate"} for the exponential distribution.}

\item{dens}{logical; if \code{TRUE} the density values are returned instead
of the cumulative distribution function.
Default is \code{FALSE}.}
}
\value{
A matrix is returned containing the values of the cumulative
distribution function (or density). Each column represents one parameter
configuration, that is, one row from the \code{parameters_values} argument.
}
\description{
Setup basis functions based on selected distribution type.
}
\details{
The argument \code{disttype} specifies the distribution type.
Anything can be used where the cumulative distribution function is available
with function name \code{pdisttype} and the argument for the
quantiles named \code{q}, and the corresponding density function as
\code{ddisttype} with argument name \code{x} for the quantile vector.
Most standard distributions in \R satisfy this condition.
}
\examples{
x <- 1:10
base_mat <- build_basis(x = x)
base_mat

}

