% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergam_methods.R
\name{check}
\alias{check}
\alias{check.ergam}
\title{Model diagnostics for ergam.}
\usage{
check(m, ...)

\method{check}{ergam}(m, print_only = TRUE, no_print = FALSE)
}
\arguments{
\item{m}{\code{\link{ergam}} object.}

\item{print_only}{logical; diagnostic information are only printed.
Default is \code{TRUE}.}
}
\value{
Nothing is returned, unless option \code{print_only = FALSE} is
specified. In this case a list is returned with elements
  \describe{
     \item{\code{which_converged}}{logical vector; indicates for which data
     subsets the algorithm did converge.}
     \item{\code{which_no_fit}}{logical vector; indicates for which data
     subsets there was no fit (due to too few observations with \eqn{y = 1}
     in the subset).}
     \item{\code{which_max_iter}}{logical vector; indicates for which data
     subsets the algorithm did reach the max. no. of iterations.}
     \item{\code{which_non_convergence}}{logical vector; indicates for which
     data subsets the algorithm did not converge.}
     \item{\code{which_effect_zero}}{logical vector; indicates for which data
     subsets at least one smooth effect was set to zero.}
     }
}
\description{
Function to summarize algorithm convergence for \code{\link{ergam}} and
print some diagnostics.
}
\examples{
\dontrun{
data("facebook")

## By default ergam function includes check functionality
mod1 <- ergam(nw = facebook, ls_num = 1, check = TRUE)

check(mod1)
}

}
\seealso{
\code{\link{ergam}}
}

