% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergam.R
\name{ergam}
\alias{ergam}
\title{Exponential Random Graph Additive Model.}
\usage{
ergam(nw, mod = c("full", "twostar", "triangle"), sig = NULL,
  disttype = "exp", parameters_names = "rate", parameters_values = NULL,
  nodes = NULL, base_num = NULL, ls_num = 0, iter_max = 20,
  y_min = 10, exact = 1e-12, use.multicore = FALSE, mc.cores = 2,
  check = TRUE, stab_fac = 1e-09, zero_tolerance = 0.005)
}
\arguments{
\item{nw}{\code{\link{network}} object or adjacency \code{\link{matrix}}.
Network has to be undirected.}

\item{mod}{character string; type of model to fit. Should be one of \cr
\code{"full"} for model with 2-star and triangle effect, \cr
\code{"twostar"} for model with only 2-star effect, or \cr
\code{"triangle"} for model with only triangle effect. \cr
Default is \code{"full"}.}

\item{sig}{integer (vector), contains the direction of the smooth effects:
\code{-1} for decreasing, \code{1} for increasing.
Two entries needed for full model (first entry for 2-star, second for
triangle effect). \cr
If not specified (default), the effect direction is determined using a GLM
fit for each data subset.}

\item{disttype}{character string; type of basis functions to use for smooth
effect(s), see details. \cr
Default is \code{"exp"} for the exponential distribution.}

\item{parameters_names}{character (vector); containing the names for the
distribution parameters. \cr
Default is \code{"rate"} for the exponential distribution.}

\item{parameters_values}{numeric vector or matrix; containing a sequence of
values for the distribution parameters. The no. of columns must be equal to
the no. of parameters. See details.}

\item{nodes}{numeric; positions where monotonicity constraint is checked,
see details.}

\item{base_num}{integer; number of basis functions to use, see details.}

\item{ls_num}{integer (vector); specifies which latin square based subsets
should be used for model fitting.\cr
The default value of \code{0} indicates all available subsets.}

\item{iter_max}{integer; maximum number of iterations, default is 20.}

\item{y_min}{integer; minimum number of observations with response 1. If a
subset contains less, fitting is skipped for this subset.}

\item{exact}{numeric; convergence criterion, default is \code{1e-12}.}

\item{use.multicore}{logical; should the algorithm run in parallel mode, if
package \link{parallel} is available? Default is \code{FALSE}.}

\item{mc.cores}{integer; no. of cores to use for parallel execution, default
is 2.}

\item{check}{logical; prints some model diagnostics using function
\code{\link{check.ergam}}. Default is \code{TRUE}.}

\item{stab_fac}{numeric; factor for avoiding numerical issues in
\code{\link{quadprog}}, default is \code{1e-9}.}

\item{zero_tolerance}{numeric; tolerance for setting effect to zero in case
of numeric difficulties. Default is \code{0.005}.}
}
\value{
an object of class \code{ergam}, for which
  \code{\link[=plot.ergam]{plot}},
  \code{\link[=check.ergam]{check}},
  \code{\link[=residuals.ergam]{residuals}}, and
  \code{\link[=gof.ergam]{gof}} methods are defined. Objects of class
  \code{ergam} are lists with entries
  \describe{
     \item{\code{results}}{a list containing the results for each of the data
     subset used, with entries \cr
     \code{alpha} for the estimated coefficients, \cr
     \code{lambda} for the estimated penalization parameter(s), \cr
     \code{status} containing information on the algorithm convergence, \cr
     \code{iter} for the number of iterations used, and \cr
     \code{sig} effect direction(s) used.}
     \item{\code{nw}}{the supplied \code{\link{network}} object.}
     \item{\code{mod}}{type of the fitted model, see argument \code{mod}.}
     \item{\code{sig}}{the direction used for the smooth effect(s) (\code{-1}
     implies decreasing, \code{1} increasing).}
     \item{\code{disttype}}{ype of basis functions used for smooth
     effect(s).}
     \item{\code{parameters_names}}{names for the distribution parameters for
     the basis functions.}
     \item{\code{parameters_values}}{values for the distribution distribution
     parameters of the basis functions.}
     \item{\code{nodes}}{positions where monotonicity constraints have been
     checked during fitting.}
     \item{\code{base_num}}{no. of basis functions used.}
     \item{\code{ls_num}}{information which subsets have been used for model
     fitting. \code{0} implies all.}
     \item{\code{iter_max}}{maximum no. of iterations.}
     \item{\code{y_min}}{minimum no. of observations with response 1 per
     subset.}
     \item{\code{exact}}{convergence criterion.}
     \item{\code{zero_tolerance}}{criterion for setting effect to zero.}
     \item{\code{max_twostar_obs}}{max. observed twostar change in data.}
     \item{\code{max_triangle_obs}}{max. observed triangle change in data.}
     }
}
\description{
Fitting an Exponential Random Graph Additive Model (ERGAM) with smooth
components.
}
\details{
The function fits an Exponential Random Graph Additive Model (ERGAM)
with smooth components to subsets of the whole network. The subsets are
chosen based on a symmetric Latin square. \cr

The argument \code{disttype} specifies the distribution type for the basis
functions of the smooth effect(s). Anything can be used where the cumulative
distribution function is available with function name \code{pdisttype} and
the argument for the quantiles named \code{q}, and the corresponding density
function as \code{disttype} with argument name \code{x} for the quantile
vector. Most standard distributions in \R satisfy this condition. \cr
If no \code{parameter_values} are sepcified and \code{"exp"} is used as
\code{disttype} (default), 20 basis functions are used for argument
\code{base_num}. In addition, for the exponential distribution the
monotonicity constraint is checked at points where neighbouring basis
functions cross each other (\code{nodes} argument). \cr

The actual fitting for \code{\link{ergam}} with smooth
twostar and/or triangle effect is done using the \code{\link{quadprog}}
routines (Turlach and Weingessel, 2013) to incorporate the
monotonicity constraints (direction as specified using \code{sig} argument or
based on automatic GLM fits per data subset).\cr
A smooth effect (twostar or triangle) can be set to zero for a data subset
during fitting (happens when the corresponding penalty parameter tends to
infinity, or if numeric issues occur in a later than the first iteration
round and the effect does not exceed the value specified in
\code{zero_tolerance}).
}
\examples{
\dontrun{
# Example akes some time to be executed
data("facebook")

mod1 <- ergam(nw = facebook, ls_num = 1:10)

plot(mod1)

plot(residuals(mod1))
}

}
\references{
Thiemichen, S. and Kauermann, G. (2016). Stable exponential random graph
models with non-parametric components for large dense networks. \cr
arXiv preprint arXiv:1604.04732. \url{http://arxiv.org/abs/1604.04732} \cr
\cr
Turlach, B. A. and Weingessel, A. (2013).
"quadprog: Functions to solve Quadratic Programming Problems". \cr
R package version 1.5-5. \url{https://CRAN.R-project.org/package=quadprog}
}

