% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facebook}
\alias{facebook}
\title{Facebook data from Stanford Large Network Dataset Collection.}
\format{An object of class \code{\link{network}} with 4039 nodes based on the
undirected binary adjacency matrix.}
\source{
\url{https://snap.stanford.edu/data/egonets-Facebook.html}
}
\usage{
data("facebook")
}
\description{
This is the combined data from 10 Facebook ego networks from Stanford Large
Network Dataset Collection. These data are stored as a undirected
\link{network} object containing the edge information only.
}
\references{
Leskovec, J. and Krevl, A. (2014) SNAP Datasets: Stanford Large
Network Dataset Collection. \url{http://snap.stanford.edu/data} \cr
McAuley, J. and Leskovec, J. (2012) Learning to Discover Social Circles in
Ego Networks. NIPS.
}
\keyword{datasets}

