% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mod.R
\name{fit_double}
\alias{fit_double}
\title{Fit full ergam model.}
\usage{
fit_double(y, C, D2, D3, A, iter_max = 20, exact = 1e-12, verbose = FALSE,
  stab_fac = 1e-09, zero_tolerance = 0.005, sig)
}
\arguments{
\item{y}{numeric vector; binary response.}

\item{C}{matrix; design matrix for the model.}

\item{D2}{matrix; penalty matrix for twostar effect.}

\item{D3}{matrix; penalty matrix for triangle effect.}

\item{A}{matrix; defines the additional monotonicity constraints taken into
account by \code{\link{quadprog}} routines.}

\item{iter_max}{integer; maximum number of iterations, default is 20.}

\item{exact}{numeric; convergence criterion, default is \code{1e-12}.}

\item{verbose}{logical; printing of current values? Default is \code{FALSE}.}

\item{stab_fac}{numeric; factor for avoiding numerical issues in
\code{\link{quadprog}}, default is \code{1e-9}.}

\item{zero_tolerance}{numeric; tolerance for setting effect to zero in case
of numeric difficulties. Default is \code{0.005}.}

\item{sig}{integer vector; giving the effect directions used.}
}
\value{
A list is returned, with elements
\describe{
 \item{alpha}{coefficient values.}
 \item{lambda}{penalty parameters (two entries: \code{lambda2} and
 \code{lambda3}).}
 \item{status}{information on algorithm status.}
 \item{iter}{number of iterations executed.}
 \item{sig}{effect directions used.}
}
}
\description{
Fit full \code{\link{ergam}} model with smooth twostar and triangle effect.
}
\details{
Function executes actual fit for \code{\link{ergam}} with smooth
twostar and triangle effect (full model) using the \code{\link{quadprog}}
routines (Turlach and Weingessel, 2013) to incorporate the additional
monotonicity constraints.\cr
Non-convergence of the algorithm or other problems during fitting can be seen
by the \code{status} entry. In addition, warnings are produced. The function
automatically checks if the maximum number of iterations was reached. \cr
If one of the two smooth effects (twostar or triangle) is set to zero
(i.e. corresponding penalty parameter tends to infinity), the function
\code{\link{fit_single}} is used accordingly to fit an \code{\link{ergam}}
model containing only the remaining smooth effect.\cr
If numeric issues occur in a later than the first iteration round, one smooth
effect (or both) is set to zero if it does not exceed the value specified in
\code{zero_tolerance}.
}
\references{
Turlach, B. A. and Weingessel, A. (2013).
"quadprog: Functions to solve Quadratic Programming Problems".
R package version 1.5-5. \url{https://CRAN.R-project.org/package=quadprog}
}

