% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mod.R
\name{fit_mod}
\alias{fit_mod}
\title{Fit ergam model.}
\usage{
fit_mod(data_sub, A, D = NULL, D2 = NULL, D3 = NULL, disttype = "exp",
  parameters_names = "rate", parameters_values = NULL, nodes = NULL,
  base_num = NULL, mod = "full", sig = NULL, iter_max = 20,
  y_min = 10, exact = 1e-12, verbose = FALSE, subset_id = FALSE,
  stab_fac = 1e-09, zero_tolerance = 0.005)
}
\arguments{
\item{data_sub}{data frame; generated using \code{\link{generate_dataset}}.}

\item{A}{matrix; defines the additional monotonicity constraints taken into
account by \code{\link{quadprog}} routines.}

\item{D}{matrix; penalty matrix in \code{"twostar"} or \code{"triangle"}
model.}

\item{D2}{matrix; penalty matrix for twostar effect in \code{"full"} model.}

\item{D3}{matrix; penalty matrix for triangle effect in \code{"full"} model.}

\item{disttype}{character string; type of basis functions to use for smooth
effects, see details. \cr
Default is \code{"exp"} for the exponential distribution.}

\item{parameters_names}{character (vector); containing the names for the
distribution parameters. \cr
Default is \code{"rate"} for the exponential distribution.}

\item{parameters_values}{numeric vector or matrix; containing a sequence of
values for the distribution parameters. The no. of columns must be equal to
the no. of parameters. See details.}

\item{nodes}{numeric; positions where monotonicity constraint is checked,
see details.}

\item{base_num}{integer; number of basis functions to use. See details.}

\item{mod}{character string; type of model to fit. Should be one of \cr
\code{"full"} for model with twostar and triangle effect, \cr
\code{"twostar"} for model with only twostar effect, or \cr
\code{"triangle"} for model with only triangle effect. \cr
Default is \code{"full"}.}

\item{sig}{integer (vector), contains the direction of the effects: \code{-1}
for decreasing, \code{1} for increasing. Two entries needed for full model
(first entry for twostar, second for triangle effect). \cr
If not specified (default), the effect direction is determined using a GLM
fit.}

\item{iter_max}{integer; maximum number of iterations, default is 20.}

\item{y_min}{integer; minimum number of observations with response 1.}

\item{exact}{numeric; convergence criterion, default is \code{1e-12}.}

\item{verbose}{logical; printing of current values? Default is \code{FALSE}.}

\item{subset_id}{logical; printing of current latin square number? Default is
\code{FALSE}.}

\item{stab_fac}{numeric; factor for avoiding numerical issues in
\code{\link{quadprog}}, default is \code{1e-9}.}

\item{zero_tolerance}{numeric; tolerance for setting effect to zero in case
of numeric difficulties. Default is \code{0.005}.}
}
\value{
A list is returned, with elements
\describe{
 \item{alpha}{coefficient values.}
 \item{lambda}{penalty parameter(s). Contains two entries (\code{lambda2} and
 \code{lambda3}) for a full model.}
 \item{status}{information on algorithm status.}
 \item{iter}{number of iterations executed.}
}
See returned objects of \code{\link{fit_double}} for a full model, and
\code{\link{fit_single}} otherwise.
}
\description{
Function to do the actual fitting of an \code{\link{ergam}} model.
}
\details{
The argument \code{disttype} specifies the distribution type for the basis
functions of the smooth effect(s). Anything can be used where the cumulative
distribution function is available with function name \code{pdisttype} and
the argument for the quantiles named \code{q}, and the corresponding density
function as \code{ddisttype} with argument name \code{x} for the quantile
vector. Most standard distributions in \R satisfy this condition. \cr
If no \code{parameter_values} are specified and \code{"exp"} is used as
\code{disttype} (default), 20 basis functions are used for argument
\code{base_num}. In addition, for the exponential distribution the
\code{nodes} argument is set so that the monotonicity constraint is checked
at points where neighbouring basis functions cross each other.\cr
If numeric issues occur in a later than the first iteration round, one smooth
effect (or both) is set to zero if it does not exceed the value specified in
\code{zero_tolerance}.
}
\seealso{
\code{\link{fit_double}}, \code{\link{fit_single}}
}

