% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate sub-dataset based on latin square.}
\usage{
generate_dataset(number = 1, ls_mat, adjacency, adjacency2,
  use.multicore = FALSE, mc.cores = 2)
}
\arguments{
\item{number}{integer vector; latin square numbers. Length is equal to the
number of returned data frames.}

\item{ls_mat}{matrix; latin square index matrix.}

\item{adjacency}{matrix; network adjacency matrix.}

\item{adjacency2}{matrix; squared network adjacency matrix.}

\item{use.multicore}{logical; if set to \code{TRUE} (and if package
\code{\link{parallel}} is available) the function is run in parallel,
default is \code{FALSE}.}

\item{mc.cores}{count; number of cores to use for parallel execution if
\code{use.multicore = FALSE}, default is 2.}
}
\value{
A list of data frames is returned, where each data frame has columns
\describe{
 \item{y}{binary response.}
 \item{edges}{edges change statictic (always 1).}
 \item{twostars}{2-star change statictic.}
 \item{triangles}{triangle change statictic.}
}
}
\description{
A function which extracts the corresponding subdataset based on a latin
square.
}
\examples{
library("network")
data("florentine", package = "ergm")

ls_mat <- simpleLS(n = get.network.attribute(flomarriage, "n"))
adj_mat <- as.matrix(flomarriage)

test_dat <- generate_dataset(number = 1,
                             ls_mat = ls_mat,
                             adjacency = adj_mat,
                             adjacency2 = adj_mat \%*\% adj_mat)
test_dat

}
\seealso{
\code{\link{ergam}}
}

