% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergam_methods.R
\name{mean.ergam}
\alias{mean.ergam}
\title{Mean estimate for ergam.}
\usage{
\method{mean}{ergam}(m, what = c("converged", "all", "nonconverged"))
}
\arguments{
\item{m}{\code{\link{ergam}} object.}

\item{what}{character string; specifies what should be used for overall
estimate calculation. Default is \code{"converged"}. See details.}
}
\value{
a named list with entries
  \describe{
     \item{\code{edges}}{numeric; the model intercept.}
     \item{\code{m_twostars}}{numeric vector; containing the values of the
     smooth functional curve for the twostar change starting at a change of
     0.}
     \item{\code{m_triangles}}{numeric vector; containing the values of the
     smooth functional curve for the triangle change starting at a change of
     0.}
  }
}
\description{
Function to get mean estimate from an \code{\link{ergam}} model.
}
\details{
The default option is to consider only converged estimates (including
estimates where smooth effects have been set to zero) for calculation of
the overall mean estimate.
If \code{what = "all"} is used, all estimated effects are used
(for non-converged estimates, the last value from the iterative
algorithm is used).
}
\seealso{
\code{\link{ergam}}.
}

