% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ergam}
\alias{plot.ergam}
\title{Plot ergam results.}
\usage{
\method{plot}{ergam}(m, what = c("converged", "all", "nonconverged"),
  xlim = NULL, plot_rug = FALSE, show_mean = TRUE, show_median = TRUE,
  alpha_shade = 0.5)
}
\arguments{
\item{m}{\code{\link{ergam}} object.}

\item{what}{character string; specifies what results should be plotted.
Default is \code{"converged"}. See details.}

\item{xlim}{numeric vector; plotting range on x-axis. If not specified,
plotting range is computed automatically.}

\item{plot_rug}{logical; should the observed values be plotted? Default is
\code{FALSE}. \cr
Advise: Do not set to \code{TRUE} for big networks.}

\item{show_mean}{logical; should a mean curve be plotted (in blue)? Default
is \code{TRUE}.}

\item{show_median}{logical; should a median curve be plotted (in orange)?
Default is \code{TRUE}.}

\item{alpha_shade}{numeric; value between 0 and 1 specifying the alpha
transparency for the plotted curve estimates.}
}
\description{
Function to plot resulting estimates for \code{\link{ergam}}.
}
\details{
The default option is to plot only converged estimates (including estimates
where smooth effects have been set to zero). If \code{what = "all"} is used,
all estimated effects are plotted (for non-converged estimates, the last
value from the iterative algorithm is used).
Option \code{what = "non-converged"} may be useful for debugging or
diagnostics if a lot of subsets encounter convergence problems. \cr
If a median curve is dislayed (\code{show_median = TRUE}) the function
\code{\link{fbplot}} from package \code{\link{fda}}
(Ramsay et. al., 2014) is used with argument \code{method = "Both"}.
The computation here relies on the fast algorithm developed by
Sun et. al. (2012).
}
\examples{
\dontrun{
data("facebook")

mod1 <- ergam(nw = facebook, ls_num = 1:2)

plot(mod1)
}

}
\references{
Ramsay, J. O., Wickham, H., Graves, S. and Hooker, G. (2014).
"fda: Functional Data Analysis". R package version 2.4.4.
\url{https://CRAN.R-project.org/package=fda} \cr\cr
Sun, Y., Genton, M. G. and Nychka, D. W. (2012),
"Exact fast computation of band depth for large functional datasets:
How quickly can one million curves be ranked?" Stat, 1(1), 68-74.
}
\seealso{
\code{\link{ergam}}
}

