% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_nw_data.R
\name{prepare_nw_data}
\alias{prepare_nw_data}
\title{Prepare network data for ergam.}
\usage{
prepare_nw_data(adjacency)
}
\arguments{
\item{adjacency}{matrix; the network adjacency matrix.}
}
\value{
A list with elements
\describe{
 \item{nnodes}{numeric; number of nodes in the network.}
 \item{max_2star_change}{numeric; maximal value for the 2-star
 change.}
 \item{max_triangle_change}{numeric; maximal value for the triangle
 change.}
 \item{adjacency}{matrix; network adjacency matrix.}
 \item{adjacency2}{matrix; squared network adjacency matrix.}
 \item{ls_mat}{matrix; corresponding latin square.}
}
}
\description{
A function to extract the needed information for the \code{\link{ergam}}
function from the whole network data set.
}
\details{
Note: If the number of nodes in the network is odd, the generated
latin square has dimension (no. of nodes - 1).
}
\examples{
library("network")
data("florentine", package = "ergm")

test_nw_dat <- prepare_nw_data(as.matrix(flomarriage))
test_nw_dat

}
\seealso{
\code{\link{ergam}}
}

