% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergam_methods.R
\name{residuals.ergam}
\alias{residuals.ergam}
\title{Residuals for ergam.}
\usage{
\method{residuals}{ergam}(m, what = c("converged", "all", "nonconverged"),
  estimate_type = c("mean", "median"))
}
\arguments{
\item{m}{\code{\link{ergam}} object.}

\item{what}{character string; specifies what should be used for overall
estimate calculation. Default is \code{"converged"}. See details.}

\item{estimate_type}{character string; specifies which estimate should be
used for residual computation, \code{"mean"} (default) or \code{"median"}
curve.}
}
\value{
an object of class \code{ergam_residuals}, for which a
  \code{\link[=plot.ergam_residuals]{plot}} method is defined.
  Objects of class
  \code{ergam_residuals} are lists with entries
  \describe{
     \item{\code{residuals}}{numeric vector; containing the mean Pearson
     residuals for each node.}
     \item{\code{what}}{character string; which results were used for overall
     estimate calculation.}
     \item{\code{estimate_type}}{character string; which estimate was used
     for goodness-of-fit diagnostics (mean or median curve).}
     }
}
\description{
Function to compute mean Pearson residuals for each node based on
\code{\link{ergam}} model.
}
\details{
An overall estimate is used to compute Pearson residuals for all in the
network adjacency matrix. For each node in the network (i.e for each row in
the adjacency) the mean value of these residuals is computed. \cr
The default option is to employ only converged estimates (including estimates
where smooth effects have been set to zero) for calculation of
the overall estimate (mean or median, see argument \code{estimate_type}).
If \code{what = "all"} is used, all estimated effects are used
(for non-converged estimates, the last value from the iterative
algorithm is used). \cr
If a median curve is used as overall estimate
(\code{estimate_type = "median"}) the function
\code{\link{fbplot}} from package \code{\link{fda}}
(Ramsay et. al., 2014) is used with argument \code{method = "Both"}.
The computation here relies on the fast algorithm developed by
Sun et. al. (2012). \cr
}
\seealso{
\code{\link{ergam}}.
}

