% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleLS.R
\name{simpleLS}
\alias{simpleLS}
\title{Simple latin square construction.}
\usage{
simpleLS(n = 10)
}
\arguments{
\item{n}{numeric; dimension of the matrix, needs to be even.}
}
\value{
a symmetric \eqn{n x n} integer matrix with 0 as diagonal elements.
}
\description{
A function which constructs a simple latin square with \eqn{n} rows and
\eqn{n} columns.
The latin square is symmetric and all diagonal entries are 0.
}
\examples{
testmat <- simpleLS(n = 10)
testmat

}
\references{
Bogomolny, A. (2016). Latin squares. Simple construction.
From Interactive Mathematics Miscellany and Puzzles
\url{http://www.cut-the-knot.org/arithmetic/latin2.shtml}.
[Online; accessed: 04 March 2016].
}

