% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_basis.R
\name{visualize_basis}
\alias{visualize_basis}
\title{Visualize basis functions of ergam.}
\usage{
visualize_basis(m, xlim = NULL)
}
\arguments{
\item{m}{\code{\link{ergam}} object.}

\item{xlim}{numeric vector; plotting range on x-axis. If not specified,
plotting range is computed automatically.}
}
\description{
Function to plot the used basis functions for an \code{\link{ergam}} fit.
}
\seealso{
\code{\link{ergam}}
}

