% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_net_to_data.R
\name{whole_net_to_data}
\alias{whole_net_to_data}
\title{Generate dataset from network.}
\usage{
whole_net_to_data(nw, range_only = FALSE)
}
\arguments{
\item{nw}{\code{\link{network}} object.}

\item{range_only}{logical; should only the \code{\link{range}} of the
variables be returned? Default is \code{FALSE}.}
}
\value{
A data frames is returned, with columns
\describe{
 \item{y}{binary response.}
 \item{edges}{edges change statictic (always 1).}
 \item{twostars}{2-star change statictic.}
 \item{triangles}{triangle change statictic.}
}
If argument \code{range_only = TRUE} the columns contain only the
\code{\link{range}} of the variables.
}
\description{
A function which extracts the information for an \code{\link{ergam}} fit
from the corresponding network.
}
\seealso{
\code{\link{ergam}}
}

